move box daily tasks logic into separate folder
