+++ src/__tests__/box/modules/box.module.ts
@@ -1,6 +1,6 @@
 import mongoose from "mongoose";
 import BoxCommonModule from "./boxCommon";
-import { ModelName } from "../../../common/enum/modelName.enum";
+import {ModelName} from "../../../common/enum/modelName.enum";
 import {BoxService} from "../../../box/box.service";
 import {BoxSchema} from "../../../box/schemas/box.schema";
 import {GroupAdminSchema} from "../../../box/groupAdmin/groupAdmin.schema";
@@ -8,40 +8,47 @@ import {GroupAdminService} from "../../../box/groupAdmin/groupAdmin.service";
 import {BoxHelper} from "../../../box/util/boxHelper";
 import BoxCreator from "../../../box/boxCreator";
 import BoxAuthHandler from "../../../box/auth/BoxAuthHandler";
+import {DailyTaskService} from "../../../box/dailyTask/dailyTask.service";
 
 export default class BoxModule {
-    private constructor() {}
+    private constructor() {
+    }
 
-    static async getBoxService(){
+    static async getBoxService() {
         const module = await BoxCommonModule.getModule();
         return module.resolve(BoxService);
     }
 
-    static getBoxModel(){
+    static getBoxModel() {
         return mongoose.model(ModelName.BOX, BoxSchema);
     }
 
-    static async getGroupAdminService(){
+    static async getGroupAdminService() {
         const module = await BoxCommonModule.getModule();
         return module.resolve(GroupAdminService);
     }
 
-    static getGroupAdminModel(){
+    static getGroupAdminModel() {
         return mongoose.model(ModelName.GROUP_ADMIN, GroupAdminSchema);
     }
 
-    static async getBoxHelper(){
+    static async getBoxHelper() {
         const module = await BoxCommonModule.getModule();
         return module.resolve(BoxHelper);
     }
 
-    static async getBoxCreator(){
+    static async getBoxCreator() {
         const module = await BoxCommonModule.getModule();
         return module.resolve(BoxCreator);
     }
 
-    static async getBoxAuthHandler(){
+    static async getBoxAuthHandler() {
         const module = await BoxCommonModule.getModule();
         return module.resolve(BoxAuthHandler);
+    }
+
+    static async getDailyTaskService() {
+        const module = await BoxCommonModule.getModule();
+        return module.resolve(DailyTaskService);
     }
 }
